+++ example/dartway_example_flutter/lib/ui_kit/3_special/auth/pin_code_text_field.dart
@@ -34,7 +34,7 @@ class PinCodeTextFieldWidget extends HookWidget {
       mainAxisAlignment: MainAxisAlignment.spaceAround,
       autoDisposeControllers: false,
       onChanged: onChanged,
-      textStyle: AppText.body.resolveStyle(context).copyWith(
+      textStyle: AppText.body.resolve(context).copyWith(
             color: context.colorScheme.onSurface,
           ),
       validator: (value) => externalError ?? internalValidator?.call(value),