+++ packages/dartway_flutter/lib/src/ui_kit/base_widgets/dw_button/dw_button_style_preset.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/material.dart';
-
-/// Contract for style presets that can be resolved into [ButtonStyle]
-/// in the given [BuildContext].
-abstract class DwButtonStylePreset {
-  const DwButtonStylePreset();
-
-  ButtonStyle resolveStyle(BuildContext context);
-}