+++ packages/dartway_flutter/lib/src/ui_kit/theme/dw_color_preset.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/material.dart';
-
-/// Contract for color presets that can be resolved into [Color]
-/// in the given [BuildContext].
-abstract class DwColorPreset {
-  const DwColorPreset();
-
-  Color resolveColor(BuildContext context);
-}